Extract ThemeData into separate module
