+++ AzooKeyCore/Sources/KeyboardThemes/ApplicationSpecificTheme.swift
@@ -1,0 +1,17 @@
+//
+//  ApplicationSpecificTheme.swift
+//
+//
+//  Created by ensan on 2023/07/20.
+//
+
+import Foundation
+import SwiftUI
+
+public protocol ApplicationSpecificTheme {
+    associatedtype ApplicationColor: ApplicationSpecificColor
+}
+
+public protocol ApplicationSpecificColor: Codable, Equatable {
+    var color: SwiftUI.Color { get }
+}
