+++ Shared/KeyboardView/View/FlickKeyboard/KeyView/FlickChangeKeyboardKeyModel.swift
@@ -47,7 +47,7 @@ struct FlickChangeKeyboardModel: FlickKeyModelProtocol {
         }
     }
 
-    func backGroundColorWhenUnpressed(states: VariableStates, theme: ThemeData) -> Color {
+    func backGroundColorWhenUnpressed(states: VariableStates, theme: AzooKeyTheme) -> Color {
         theme.specialKeyFillColor.color
     }
